Change how we calculate the seed
Instead of using my handrolled hashing system, I'm using SHA256
instead to generate a seed for the generator. Doing it this way
reduces the probablity of hash collisions.

In the past, collisions was considered an intended feature of sorts,
since it allow people to generate the same mission even if they make
a small typo. However, the existence of hash collisions also affects
```railroad.rb```, as it reduces the "effective" search range of
that program. We would sometimes stumble upon the same mission
when re-running ```railroad.rb``` with the same search term -
and the only difference would just be the word we used to generate
the text.

I also noticed that I don't actually *care* what I named
the mission, so typos aren't really a big issue to worry about.

So, we switched from my old system over to SHA256 to make
```railroad.rb``` more robust. This will wipe out all existing missions,
though.